Move all feature pages and layouts to components folder, and update routes imports
