+++ frontend/src/features/account/components/AccountSetupPage.test.tsx
@@ -1,10 +1,10 @@
 import userEvent from "@testing-library/user-event";
 import { describe, expect, test, vi } from "vitest";
 
-import { AccountSetupPage } from "@/features/account";
-
 import { AccountUpdateRequestHandler, loginResponseMockData } from "@kiesraad/api-mocks";
 import { render, screen, server } from "@kiesraad/test";
+
+import { AccountSetupPage } from "./AccountSetupPage";
 
 const navigate = vi.fn();
