+++ frontend/src/features/account/components/Logout.tsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { Navigate } from "react-router";
 
-import { useApiState } from "@kiesraad/api";
+import { useApiState } from "@/api";
 
 export function Logout() {
   const { logout } = useApiState();
