--- frontend/src/features/account/index.ts
@@ -1,2 +1,0 @@
-export * from "./LoginLayout";
-export * from "./page";
