--- frontend/src/features/account/page/index.ts
@@ -1,3 +1,0 @@
-export * from "./LoginPage";
-export * from "./AccountSetupPage";
-export * from "./UserHomePage";
