+++ frontend/src/features/data_entry/components/CandidatesVotesPage.tsx
