+++ frontend/src/features/data_entry/components/DataEntryHomePage.tsx
