+++ frontend/src/features/data_entry/components/DifferencesPage.tsx
