+++ frontend/src/features/data_entry/components/VotersAndVotesPage.tsx
