--- frontend/src/features/data_entry/index.ts
@@ -1,1 +1,0 @@
-export * from "./page";
