--- frontend/src/features/data_entry/page/index.ts
@@ -1,5 +1,0 @@
-export * from "./CandidatesVotesPage";
-export * from "./DataEntryHomePage";
-export * from "./DifferencesPage";
-export * from "./RecountedPage";
-export * from "./VotersAndVotesPage";
