+++ frontend/src/features/election/components/ElectionHomePage.tsx
