+++ frontend/src/features/election/components/ElectionReportPage.tsx
