+++ frontend/src/features/election/components/ElectionStatusPage.tsx
