+++ frontend/src/features/election/components/OverviewLayout.tsx
@@ -1,6 +1,7 @@
 import { Outlet } from "react-router";
 
-import { ElectionListProvider } from "@kiesraad/api";
+import { ElectionListProvider } from "@/api";
+
 import { AppLayout } from "@kiesraad/ui";
 
 export function OverviewLayout() {
