--- frontend/src/features/election/index.ts
@@ -1,3 +1,0 @@
-export * from "./ElectionLayout";
-export * from "./OverviewLayout";
-export * from "./page";
