--- frontend/src/features/election/page/index.ts
@@ -1,4 +1,0 @@
-export * from "./ElectionHomePage";
-export * from "./ElectionReportPage";
-export * from "./ElectionStatusPage";
-export * from "./OverviewPage";
