--- frontend/src/features/logs/index.ts
@@ -1,1 +1,0 @@
-export * from "./LogsHomePage";
