+++ frontend/src/features/polling_stations/components/PollingStationDeleteModal.tsx
