+++ frontend/src/features/polling_stations/components/PollingStationListPage.tsx
