+++ frontend/src/features/polling_stations/components/PollingStationsLayout.tsx
