--- frontend/src/features/polling_stations/index.ts
@@ -1,2 +1,0 @@
-export * from "./PollingStationsLayout";
-export * from "./page";
