--- frontend/src/features/polling_stations/page/index.ts
@@ -1,1 +1,0 @@
-export * from "./PollingStationListPage";
