+++ frontend/src/features/workstations/components/WorkstationsHomePage.tsx
@@ -1,4 +1,5 @@
-import { t } from "@kiesraad/i18n";
+import { t } from "@/lib/i18n";
+
 import { PageTitle } from "@kiesraad/ui";
 
 export function WorkstationsHomePage() {
