--- frontend/src/features/workstations/index.ts
@@ -1,1 +1,0 @@
-export * from "./WorkstationsHomePage";
